backend/requirements.txt
django==4.1
djangorestframework
djangorestframework-jsonapi
pymysql
Pillow
